requirements.txt
selenium
requests
pytest
allure-pytest
pytest-xdist
pytest-rerunfailures